refactor: reorganize config package
